Adding 'env' function to print environment.
Slightly odd semantics because language doesn't allow for zero argument
functions. I opted for env taking a list, which if empty returns all
variables, but if symbols are passed, prints out those symbols.

I had also implemented a zero argument version, but that limits the
ability for passing functions around - not sure if this will be useful
later or not, so I've opted for this implementation.