Adding an exit function.
This has slightly different semantics than just hitting CTRL-D (which
ends the REPL and cleanly exits).

This exit function evaluates it's parameters then immediately exits
without any cleanup. The idea here is that it could be used like an
abort or exception in our language when we just want to bail as fast as
possible.