Initial lambda/call success.
Added lambda type. Separating it from builtins because I'm still using
a union and it made testing for builtin being non-null a no-go.

I renamed builtin_var to lenv_add because FUCK ME does this author
suffers from poor naming skills. I also made it take a global boolean
instead of a string.

I'm trying not to freak out on this whole thing... but this design is
trying my patience.